# src/backend/docs/swagger.yaml
basePath: /api/v1/nlip
definitions:
...
  token.CreateTokenRequest:
    properties:
      description:
        maxLength: 100
        type: string
      expiryDays:
        type: integer
    required:
    - description
    type: object
  token.CreateTokenResponse:
    properties:
      token:
        type: string
      tokenInfo:
        $ref: '#/definitions/token.Token'
    type: object
  token.ListTokensResponse:
    properties:
      maxItems:
        type: integer
      tokens:
        items:
          $ref: '#/definitions/token.Token'
        type: array
    type: object
  token.Token:
    properties:
      createdAt:
        type: string
      description:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      lastUsedAt:
        type: string
      token:
        type: string
      userId:
        type: string
    type: object
  token.TokenLoginRequest:
    properties:
      token:
        type: string
      username:
        type: string
    required:
    - token
    - username
    type: object
  token.TokenLoginResponse:
    properties:
      jwtToken:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  user.AuthResponse:
    properties:
      needChangePwd:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  user.ChangePasswordRequest:
    properties:
      newPassword:
        maxLength: 50
        minLength: 6
        type: string
      oldPassword:
        maxLength: 50
        minLength: 6
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
  user.GetCurrentUserResponse:
    properties:
      id:
        type: string
      isAdmin:
        type: boolean
      username:
        type: string
    type: object
  user.LoginRequest:
    properties:
      password:
        maxLength: 50
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.RegisterRequest:
    properties:
      password:
        maxLength: 50
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
  user.User:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isAdmin:
        type: boolean
      needChangePwd:
        type: boolean
      username:
        type: string
    type: object
info:
  contact: {}
  description: Nlip API
  title: Nlip API
  version: "1.0"
paths:
  /api/v1/nlip/auth/change-password:
    post:
      consumes:
      - application/json
      description: 修改当前用户的密码
      parameters:
      - description: 修改密码请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            type: string
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "403":
          description: 旧密码错误
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - 认证
  /api/v1/nlip/auth/login:
    post:
      consumes:
      - application/json
      description: 使用用户名和密码进行登录
      parameters:
      - description: 登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 用户名或密码错误
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      summary: 用户登录
      tags:
      - 认证
  /api/v1/nlip/auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的基本信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取用户信息成功
          schema:
            $ref: '#/definitions/user.GetCurrentUserResponse'
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /api/v1/nlip/auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 注册请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/user.AuthResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "409":
          description: 用户名已存在
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      summary: 用户注册
      tags:
      - 认证
  /api/v1/nlip/auth/token-login:
    post:
      consumes:
      - application/json
      description: 使用预先生成的Token进行登录
      parameters:
      - description: Token登录请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/token.TokenLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/token.TokenLoginResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: Token无效或已过期
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      summary: Token登录
      tags:
      - 认证
...
  /api/v1/nlip/tokens:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有Token列表
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/token.ListTokensResponse'
        "401":
          description: 未授权
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Token列表
      tags:
      - Token
    post:
      consumes:
      - application/json
      description: 为用户创建新的访问Token，可用于API认证
      parameters:
      - description: 创建Token请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/token.CreateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/token.CreateTokenResponse'
        "400":
          description: 请求参数错误
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "403":
          description: 达到最大Token数量限制
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 创建Token
      tags:
      - Token
  /api/v1/nlip/tokens/{id}:
    delete:
      consumes:
      - application/json
      description: 撤销指定的Token，使其失效
      parameters:
      - description: 要撤销的Token ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 撤销成功
          schema:
            type: string
        "401":
          description: 未授权
          schema:
            type: string
        "404":
          description: Token不存在
          schema:
            type: string
        "500":
          description: 服务器内部错误
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 撤销Token
      tags:
      - Token
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"